backend/requirements.txt
fastapi
httpx
python-dotenv
supabase
python-jose[cryptography]
python-multipart
email-validator
aiohttp 